Factor out helper and unit tests
