+++ src/charm.py
@@ -4,63 +4,6 @@
 
 """Charmed Machine Operator for the PostgreSQL database."""
 
-
-def _remove_stale_otel_sdk_packages():
-    """Hack to remove stale opentelemetry sdk packages from the charm's python venv.
-
-    See https://github.com/canonical/grafana-agent-operator/issues/146 and
-    https://bugs.launchpad.net/juju/+bug/2058335 for more context. This patch can be removed after
-    this juju issue is resolved and sufficient time has passed to expect most users of this library
-    have migrated to the patched version of juju.  When this patch is removed, un-ignore rule E402 for this file in the pyproject.toml (see setting
-    [tool.ruff.lint.per-file-ignores] in pyproject.toml).
-
-    This only has an effect if executed on an upgrade-charm event.
-    """
-    # all imports are local to keep this function standalone, side-effect-free, and easy to revert later
-    import os
-
-    if os.getenv("JUJU_DISPATCH_PATH") != "hooks/upgrade-charm":
-        return
-
-    import logging
-    import shutil
-    from collections import defaultdict
-
-    from importlib_metadata import distributions
-
-    otel_logger = logging.getLogger("charm_tracing_otel_patcher")
-    otel_logger.debug("Applying _remove_stale_otel_sdk_packages patch on charm upgrade")
-    # group by name all distributions starting with "opentelemetry_"
-    otel_distributions = defaultdict(list)
-    for distribution in distributions():
-        name = distribution._normalized_name
-        if name.startswith("opentelemetry_"):
-            otel_distributions[name].append(distribution)
-
-    otel_logger.debug(f"Found {len(otel_distributions)} opentelemetry distributions")
-
-    # If we have multiple distributions with the same name, remove any that have 0 associated files
-    for name, distributions_ in otel_distributions.items():
-        if len(distributions_) <= 1:
-            continue
-
-        otel_logger.debug(f"Package {name} has multiple ({len(distributions_)}) distributions.")
-        for distribution in distributions_:
-            if not distribution.files:  # Not None or empty list
-                path = distribution._path
-                otel_logger.info(f"Removing empty distribution of {name} at {path}.")
-                shutil.rmtree(path)
-
-    otel_logger.debug("Successfully applied _remove_stale_otel_sdk_packages patch. ")
-
-
-# apply hacky patch to remove stale opentelemetry sdk packages on upgrade-charm.
-# it could be trouble if someone ever decides to implement their own tracer parallel to
-# ours and before the charm has inited. We assume they won't.
-# !!IMPORTANT!! keep all otlp imports UNDER this call.
-_remove_stale_otel_sdk_packages()
-
-# ruff: disable[E402]
 import json
 import logging
 import os
@@ -76,6 +19,15 @@ from pathlib import Path
 from typing import Literal, get_args
 from urllib.parse import urlparse
 
+from utils import _remove_stale_otel_sdk_packages
+
+# apply hacky patch to remove stale opentelemetry sdk packages on upgrade-charm.
+# it could be trouble if someone ever decides to implement their own tracer parallel to
+# ours and before the charm has inited. We assume they won't.
+# !!IMPORTANT!! keep all otlp imports UNDER this call.
+_remove_stale_otel_sdk_packages()
+
+# ruff: disable[E402]
 import psycopg2
 from charmlibs import snap
 from charms.data_platform_libs.v0.data_interfaces import DataPeerData, DataPeerUnitData
