+++ tests/unit/test_utils.py
@@ -2,10 +2,10 @@
 # See LICENSE file for licensing details.
 
 import re
-from unittest.mock import patch
+from unittest.mock import Mock, patch, sentinel
 
 from constants import POSTGRESQL_SNAP_NAME
-from utils import new_password, snap_refreshed
+from utils import _remove_stale_otel_sdk_packages, new_password, snap_refreshed
 
 
 def test_new_password():
@@ -34,3 +34,47 @@ def test_snap_refreshed():
     ):
         assert snap_refreshed("100") is False
         assert snap_refreshed("200") is False
+
+
+def test_remove_stale_otel_sdk_packages():
+    with (
+        patch("utils.os.getenv", return_value=None) as _getenv,
+        patch("utils.shutil") as _shutil,
+        patch("utils.distributions") as _distributions,
+    ):
+        other_dist = Mock()
+        other_dist._normalized_name = "test"
+        otel_dist = Mock()
+        otel_dist._normalized_name = "opentelemetry_test"
+        stale_otel_dist = Mock()
+        stale_otel_dist._normalized_name = "opentelemetry_test"
+        stale_otel_dist.files = []
+        stale_otel_dist._path = sentinel.path
+
+        # Not called if not upgrade hook
+        _remove_stale_otel_sdk_packages()
+
+        _distributions.assert_not_called()
+        _shutil.rmtree.assert_not_called()
+        _distributions.reset_mock()
+        _shutil.rmtree.reset_mock()
+
+        # Upgrade hook, nothing to remove
+        _getenv.return_value = "hooks/upgrade-charm"
+        _distributions.return_value = [other_dist, otel_dist]
+
+        _remove_stale_otel_sdk_packages()
+
+        _distributions.assert_called_once_with()
+        _shutil.rmtree.assert_not_called()
+        _distributions.reset_mock()
+        _shutil.rmtree.reset_mock()
+
+        # Upgrade hook, duplicate otel packages
+        _distributions.return_value = [other_dist, otel_dist, stale_otel_dist]
+        _remove_stale_otel_sdk_packages()
+
+        _distributions.assert_called_once_with()
+        _shutil.rmtree.assert_called_once_with(sentinel.path)
+        _distributions.reset_mock()
+        _shutil.rmtree.reset_mock()
